pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atams"
version = "0.1.0"
description = "Advanced Toolkit for Application Management System"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pycryptodome>=3.19.0",
    "httpx>=0.25.0",
]
